feat(Navigation Component) added navigation component and did some changes that depend from navigation componen (Fragments, FragmentModule, etc.).
