+++ app/src/main/java/com/example/testapplication/di/module/ActivityModule.kt
@@ -1,15 +1,11 @@
 package com.example.testapplication.di.module
 
-import com.example.testapplication.ui.post.PostActivity
-import com.example.testapplication.ui.user.UserActivity
+import com.example.testapplication.ui.main.MainActivity
 import dagger.Module
 import dagger.android.ContributesAndroidInjector
 
 @Module
 abstract class ActivityModule {
     @ContributesAndroidInjector
-    abstract fun contributeUserActivity(): UserActivity
-
-    @ContributesAndroidInjector
-    abstract fun contributePostActivity(): PostActivity
+    abstract fun contributeMainActivity(): MainActivity
 }
