+++ app/src/main/java/com/example/testapplication/di/module/FragmentModule.kt
@@ -1,0 +1,15 @@
+package com.example.testapplication.di.module
+
+import com.example.testapplication.ui.post.PostFragment
+import com.example.testapplication.ui.user.UserFragment
+import dagger.Module
+import dagger.android.ContributesAndroidInjector
+
+@Module
+abstract class FragmentModule {
+    @ContributesAndroidInjector
+    abstract fun contributeUserFragment(): UserFragment
+
+    @ContributesAndroidInjector
+    abstract fun contributePostFragment(): PostFragment
+}
