+++ app/src/main/java/com/example/testapplication/ui/main/MainActivity.kt
@@ -1,0 +1,37 @@
+package com.example.testapplication.ui.main
+
+import android.os.Bundle
+import android.view.View
+import androidx.appcompat.app.AppCompatDelegate
+import androidx.databinding.DataBindingUtil
+import androidx.navigation.findNavController
+import com.example.testapplication.R
+import com.example.testapplication.databinding.ActivityMainBinding
+import com.example.testapplication.ui.user.UserFragmentDirections
+import com.google.android.material.card.MaterialCardView
+import dagger.android.support.DaggerAppCompatActivity
+
+class MainActivity : DaggerAppCompatActivity() {
+    private lateinit var binding: ActivityMainBinding
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        binding = DataBindingUtil.setContentView(this, R.layout.activity_main)
+        setSupportActionBar(binding.appBar.toolbar)
+    }
+
+    fun openPostsOfUser(v: View) {
+        val id = Integer.parseInt((v as MaterialCardView).contentDescription.toString())
+        val action = UserFragmentDirections.actionUserFragmentToPostFragment(id)
+        v.findNavController().navigate(action)
+    }
+    fun changeTheme(v: View) {
+        val nightMode = AppCompatDelegate.getDefaultNightMode()
+        if (nightMode == AppCompatDelegate.MODE_NIGHT_YES) {
+            AppCompatDelegate.setDefaultNightMode(AppCompatDelegate.MODE_NIGHT_NO)
+        } else {
+            AppCompatDelegate.setDefaultNightMode(AppCompatDelegate.MODE_NIGHT_YES)
+        }
+        recreate()
+    }
+}
