--- app/src/main/res/layout/activity_user.xml
@@ -1,20 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<layout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools">
-
-    <data></data>
-
-    <FrameLayout
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        tools:context=".ui.user.UserActivity">
-
-        <include
-            android:id="@+id/app_bar"
-            layout="@layout/app_bar" />
-
-        <include
-            android:id="@+id/content_user"
-            layout="@layout/content_users" />
-    </FrameLayout>
-</layout>
