+++ app/src/main/res/layout/app_bar.xml
@@ -18,25 +18,13 @@
         android:layout_width="55dp"
         android:layout_height="0dp"
         android:src="@drawable/ic_sun"
+        android:background="@null"
         android:onClick="changeTheme"
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toTopOf="parent" />
 
-    <com.google.android.material.textfield.TextInputLayout
-        android:id="@+id/textInputLayout"
-        android:layout_width="252dp"
-        android:layout_height="61dp"
-        app:layout_constraintStart_toEndOf="@+id/btnTheme"
-        app:layout_constraintTop_toTopOf="parent">
 
-        <com.google.android.material.textfield.TextInputEditText
-            android:id="@+id/tvSearchTitle"
-            android:layout_width="match_parent"
-            android:layout_height="match_parent"
-            android:hint="@string/title" />
-
-    </com.google.android.material.textfield.TextInputLayout>
 
     <androidx.appcompat.widget.Toolbar
         android:id="@+id/toolbar"
@@ -44,18 +32,8 @@
         android:layout_height="match_parent"
         android:layout_marginStart="50dp"
         app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toEndOf="@+id/textInputLayout"
         app:layout_constraintTop_toTopOf="@+id/btnTheme" />
 
-    <ImageButton
-        android:id="@+id/btnSearchTitle"
-        android:layout_width="49dp"
-        android:layout_height="52dp"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toStartOf="@+id/toolbar"
-        app:layout_constraintStart_toEndOf="@+id/textInputLayout"
-        app:layout_constraintTop_toTopOf="parent"
-        app:srcCompat="@android:drawable/ic_menu_search" />
 </androidx.constraintlayout.widget.ConstraintLayout>
     </com.google.android.material.appbar.AppBarLayout>
 </layout>
