+++ app/src/main/res/layout/item_comment.xml
@@ -21,7 +21,7 @@
             android:elevation="8dp"
             android:padding="15dp"
             app:cardCornerRadius="20dp"
-            app:cardElevation="8dp">
+            app:cardElevation="10dp">
 
             <androidx.constraintlayout.widget.ConstraintLayout
                 android:layout_width="match_parent"
