+++ app/src/main/res/values/strings.xml
@@ -6,4 +6,6 @@
     <string name="address">address</string>
     <string name="title">title</string>
     <string name="body">body</string>
+    <!-- TODO: Remove or change this placeholder text -->
+    <string name="hello_blank_fragment">Hello blank fragment</string>
 </resources>
